src/test/java/dev/kmunton/year2023/day23/Day23Test.java
package dev.kmunton.year2023.day23;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.util.List;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;

class Day23Test {

  private static Day23 day;
  private static final List<String> INPUT = """
      #.#####################
      #.......#########...###
      #######.#########.#.###
      ###.....#.>.>.###.#.###
      ###v#####.#v#.###.#.###
      ###.>...#.#.#.....#...#
      ###v###.#.#.#########.#
      ###...#.#.#.......#...#
      #####.#.#.#######.#.###
      #.....#.#.#.......#...#
      #.#####.#.#.#########v#
      #.#...#...#...###...>.#
      #.#.#v#######v###.###v#
      #...#.>.#...>.>.#.###.#
      #####v#.#.###v#.#.###.#
      #.....#...#...#.#.#...#
      #.#########.###.#.#.###
      #...###...#...#...#.###
      ###.###.#.###v#####v###
      #...#...#.#.>.>.#.>.###
      #.###.###.#.###.#.#v###
      #.....###...###...#...#
      #####################.#""".lines().toList();

  @BeforeAll
  static void beforeAll() {
    day = new Day23(INPUT);
  }

  @Test
  void part1() {
    assertEquals(94, day.part1());

  }

  @Test
  void part2() {
    assertEquals(154, day.part2());
  }
}
